Core/DataStores: Removed overriding field sign type based on db2 column compression
(cherry picked from commit 89f490bdcdf88c2576c17218926e89b1082ae682)

# Conflicts:
#	sql/updates/auth/cata_classic/2024_09_27_00_hotfixes.sql
#	src/server/game/Achievements/CriteriaHandler.cpp
#	src/server/game/DataStores/DB2LoadInfo.h
#	src/server/game/DataStores/DB2Stores.cpp
#	src/server/game/DataStores/DB2Stores.h
#	src/server/game/DataStores/DB2Structure.h
#	src/server/game/DataStores/DBCEnums.h
#	src/server/game/Entities/Item/AzeriteItem/AzeriteEmpoweredItem.cpp
#	src/server/game/Entities/Item/ItemBonusMgr.cpp
#	src/server/game/Entities/Player/Player.cpp
#	src/server/game/Spells/TraitMgr.cpp

+++ src/server/game/Achievements/CriteriaHandler.cpp
@@ -710,7 +710,7 @@ void CriteriaHandler::UpdateCriteria(CriteriaType type, uint64 miscValue1 /*= 0*
                     SkillLineAbilityMapBounds bounds = sSpellMgr->GetSkillLineAbilityMapBounds(spellId);
                     for (SkillLineAbilityMap::const_iterator skillIter = bounds.first; skillIter != bounds.second; ++skillIter)
                     {
-                        if (skillIter->second->SkillLine == int32(criteria->Entry->Asset.SkillID))
+                        if (skillIter->second->SkillLine == uint32(criteria->Entry->Asset.SkillID))
                         {
                             // do not add couter twice if by any chance skill is listed twice in dbc (eg. skill 777 and spell 22717)
                             ++spellCount;
@@ -3347,7 +3347,7 @@ bool CriteriaHandler::ModifierSatisfied(ModifierTreeEntry const* modifier, uint6
                 if (item->GetVisibleAppearanceModId(referencePlayer) == itemModifiedAppearance->ID)
                     return ItemSearchCallbackResult::Stop;
 
-                if (int32(item->GetEntry()) == itemModifiedAppearance->ItemID)
+                if (item->GetEntry() == itemModifiedAppearance->ItemID)
                     return ItemSearchCallbackResult::Stop;
 
                 return ItemSearchCallbackResult::Continue;
